Added feature in Listener.java to ignore files and folders that begin with a "." when overriding.  This way, if you check your customized template folder into SVN, the .svn files/folders won't cause it to break.
git-svn-id: http://svn.openmrs.org/branches/alpha@917 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ src/web/org/openmrs/web/Listener.java
@@ -34,7 +34,7 @@ public final class Listener implements ServletContextListener {
 	 */
 	public void contextInitialized(ServletContextEvent event) {
 		Log log = LogFactory.getLog(this.getClass());
-
+		
 		try {
 			FileInputStream propertyStream = null;
 
@@ -81,27 +81,27 @@ public final class Listener implements ServletContextListener {
 				if (userOverridePath != null) {
 					String absolutePath = realPath + webappPath;
 					File file = new File(userOverridePath);
+
 					// if they got the path correct
-					if (file.exists()) {
+					// also, if file does not start with a "." (hidden files, like SVN files) 
+					if (file.exists() && !userOverridePath.startsWith(".")) {
 						log.debug("Overriding file: " + absolutePath);
 						log.debug("Overriding file with: " + userOverridePath);
 						if (file.isDirectory()) {
 							for (File f : file.listFiles()) {
 								userOverridePath = f.getAbsolutePath();
-								String tmpAbsolutePath = absolutePath + "/"
-										+ f.getName();
-								FileInputStream inputStream = new FileInputStream(
-										userOverridePath);
-								FileOutputStream outputStream = new FileOutputStream(
-										tmpAbsolutePath);
-								Helper.copyFile(inputStream, outputStream);
+								if ( !f.getName().startsWith(".") ) {
+									String tmpAbsolutePath = absolutePath + "/"
+									+ f.getName();
+									FileInputStream inputStream = new FileInputStream(userOverridePath);
+									FileOutputStream outputStream = new FileOutputStream(tmpAbsolutePath);
+									Helper.copyFile(inputStream, outputStream);
+								}
 							}
 						} else {
 							// file is not a directory
-							FileInputStream inputStream = new FileInputStream(
-									userOverridePath);
-							FileOutputStream outputStream = new FileOutputStream(
-									absolutePath);
+							FileInputStream inputStream = new FileInputStream(userOverridePath);
+							FileOutputStream outputStream = new FileOutputStream(absolutePath);
 							Helper.copyFile(inputStream, outputStream);
 						}
 					}
@@ -119,7 +119,6 @@ public final class Listener implements ServletContextListener {
 			FormEntryUtil.startup();
 			SchedulerUtil.startup();
 		}
-
 	}
 
 	public void contextDestroyed(ServletContextEvent event) {
